pyproject.toml
[project]
name = "crm-extender-poc"
version = "0.1.0"
description = "Gmail Conversation Aggregation PoC"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "anthropic>=0.39.0",
    "rich>=13.7.0",
    "mail-parser-reply>=0.1.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
poc = "poc.__main__:main"

[tool.setuptools.packages.find]
include = ["poc*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
